pyproject.toml
[tool.poetry]
name = "djangotutorial"
version = "0.1.0"
description = ""
authors = ["Сергей Коновалов <SokiZNosoK68rus@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
django-debug-toolbar = "^4.4.6"
setuptools = "^75.6.0"
django-polls = {path = "/Users/sergey/Desktop/django-polls/dist/django-polls-0.1.tar.gz"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
